% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratings.R
\name{compute_ratings}
\alias{compute_ratings}
\title{Compute Player Ratings}
\usage{
compute_ratings(goals, assists, passes, tackles, saves, shots)
}
\arguments{
\item{goals}{Number of goals scored.}

\item{assists}{Number of assists.}

\item{passes}{Number of successful passes.}

\item{tackles}{Number of tackles won.}

\item{saves}{Number of saves (for goalkeepers).}

\item{shots}{Number of shots taken.}
}
\value{
A numeric rating between 0 and 10.
}
\description{
This function calculates a player's rating based on performance stats.
}
